+++ packages/terra-dropdown-button/tests/wdio/dropdown-button-spec.js
@@ -21,9 +21,14 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       $('[class*=dropdown-button]').click();
       Terra.validates.element('dropdown open', { selector: '[class*=dropdown-list]' });
       Terra.validates.element('dropdown button with dropdown open');
+      // to validate focus on button when dropdown is closed by outside click
+      $('#root').click();
+      expect($('[class*=dropdown-button]')).toBeFocused();
     });
 
     it('should print MetaData of option on click', () => {
+      // added this event to open dropdown again
+      $('[class*=dropdown-button]').click();
       $('#opt1').click();
       Terra.validates.element('MetaData of 1st option', { selector: '#root' });
     });
@@ -90,17 +95,17 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       });
 
       it('should run callback on space', () => {
-        browser.keys(['ArrowDown', 'Space']);
+        browser.keys(['Space']);
         Terra.validates.element('space');
       });
 
       it('keyboard navigates down and runs callback on space', () => {
-        browser.keys(['ArrowDown', 'ArrowDown', 'Space']);
+        browser.keys(['ArrowDown', 'Space']);
         Terra.validates.element('down arrow');
       });
 
       it('should run callback on enter', () => {
-        browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'Enter']);
+        browser.keys(['ArrowDown', 'ArrowDown', 'Enter']);
         Terra.validates.element('enter');
       });
 
@@ -110,14 +115,14 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       });
 
       it('Navigates from last item to first item', () => {
-        browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowDown']);
+        browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown']);
         expect($('#red').isFocused()).toBeTruthy();
         browser.keys(['Enter']);
         Terra.validates.element('first item focused');
       });
 
       it('keyboard navigates up and runs callback on enter', () => {
-        browser.keys(['ArrowDown', 'ArrowDown', 'ArrowUp', 'Enter']);
+        browser.keys(['ArrowDown', 'ArrowUp', 'Enter']);
         Terra.validates.element('up arrow');
       });
 
@@ -129,17 +134,17 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       });
 
       it('jumps to the last entry', () => {
-        browser.keys(['ArrowDown', 'End', 'Enter']);
+        browser.keys(['End', 'Enter']);
         Terra.validates.element('end');
       });
 
       it('jumps to the first entry', () => {
-        browser.keys(['ArrowDown', 'Home', 'Enter']);
+        browser.keys(['Home', 'Enter']);
         Terra.validates.element('home');
       });
 
       it('jumps when typing', () => {
-        browser.keys(['ArrowDown', 'b', 'Enter']);
+        browser.keys(['b', 'Enter']);
         Terra.validates.element('jumps when typing');
       });
 
@@ -267,7 +272,6 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
     it('Focus to single item on arrow down', () => {
       $('[class*=dropdown-button]').click();
       $('[class*=dropdown-list]').waitForDisplayed();
-      browser.keys(['ArrowDown', 'ArrowDown']);
       Terra.validates.element('single item focus', { selector: '#root' });
     });
   });