+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added documentation updates for `terra-form-input
+  * Added documentation updates for `terra-form-input`.
 
 ## 1.31.0 - (July 13, 2023)