+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation updates for `terra-form-input`.
+
 ## 1.31.0 - (July 13, 2023)
 
 * Added